Rename ModelTab and CoercedQueryTab to InternalLink::Model and InternalLink::CoercedQuery
